refactor: mudando average para double

+++ src/main/java/dev/valente/desafiovagaitau/dto/StatisticsResponseDTO.java
@@ -5,5 +5,5 @@ import lombok.Builder;
 import java.math.BigDecimal;
 
 @Builder
-public record StatisticsResponseDTO(double sum, BigDecimal average, double min, double max, long count) {
+public record StatisticsResponseDTO(double sum, double average, double min, double max, long count) {
 }
